Initial commit
I came up with 2 solutions for the Tennis Kata
Both solutions are similar.
In the first solution It bugged me that the GameOver is represented by a special case of S tuple (i.e: (S.Win, S.Loose)). I Wanted to have a more mature modeling of the game state. That's why I came up with the 2nd solution